Remove GoogleAPIHelper
