+++ QuizFight/app/src/main/java/rogueone/quizfight/SignInActivity.java
@@ -9,7 +9,6 @@ import retrofit2.converter.gson.GsonConverterFactory;
 import rogueone.quizfight.rest.EndpointInterface;
 import rogueone.quizfight.models.User;
 import rogueone.quizfight.utils.BaseGameUtils;
-import rogueone.quizfight.utils.GoogleAPIHelper;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -28,12 +27,13 @@ import com.google.firebase.iid.FirebaseInstanceId;
  * Created by mdipirro on 19/05/17.
  */
 
-public class SignInActivity extends AppCompatActivity {
+public class SignInActivity extends AppCompatActivity implements
+        GoogleApiClient.ConnectionCallbacks, GoogleApiClient.OnConnectionFailedListener {
 
     private static final String TAG     = "SignInActivity";
     private static final int RC_SIGN_IN = 9001;
 
-    private GoogleAPIHelper client;
+    private GoogleApiClient client;
 
     private boolean resolvingConnectionFailure  = false;
     private boolean autoStartSignInFlow         = true;
@@ -46,22 +46,24 @@ public class SignInActivity extends AppCompatActivity {
 
         setContentView(R.layout.activity_sign_in);
 
-        client = ((QuizFightApplication)getApplicationContext()).getGoogleAPIHelper();
+        client = new GoogleApiClient.Builder(this)
+                .addConnectionCallbacks(this)
+                .addOnConnectionFailedListener(this)
+                .addApi(Games.API).addScope(Games.SCOPE_GAMES)
+                .build();
+
+        ((QuizFightApplication)getApplicationContext()).setClient(client);
     }
 
     public void signIn(View v) {
-        signInClicked = true;
-        inSignInFlow = true;
         client.connect();
     }
 
     @Override
     protected void onStart() {
         super.onStart();
-
         if (!inSignInFlow) {
             // auto sign in
-            inSignInFlow = true;
             client.connect();
         }
     }
@@ -69,7 +71,49 @@ public class SignInActivity extends AppCompatActivity {
     @Override
     protected void onActivityResult(int requestCode, int resultCode, Intent data) {
         if (resultCode == RESULT_OK) {
-            client.connect();
+            Intent intent = new Intent(SignInActivity.this, HomeActivity.class);
+            intent.setFlags(Intent.FLAG_ACTIVITY_CLEAR_TOP);
+            startActivity(intent);
         }
+    }
+
+    @Override
+    public void onConnected(Bundle bundle) {
+        Retrofit retrofit = new Retrofit.Builder()
+                .baseUrl(getString(R.string.base_url))
+                .addConverterFactory(GsonConverterFactory.create())
+                .build();
+        EndpointInterface apiService = retrofit.create(EndpointInterface.class);
+        Call<ResponseBody> addToken = apiService.addToken(new User(
+                Games.getCurrentAccountName(client),
+                FirebaseInstanceId.getInstance().getToken().toString()
+        ));
+        addToken.enqueue(new Callback<ResponseBody>() {
+            @Override
+            public void onResponse(Call<ResponseBody> call, Response<ResponseBody> response) {
+                /*Intent intent = new Intent(context, HomeActivity.class);
+                context.startActivity(intent);*/
+                Intent intent = new Intent(SignInActivity.this, HomeActivity.class);
+                intent.setFlags(Intent.FLAG_ACTIVITY_REORDER_TO_FRONT);
+                startActivityIfNeeded(intent, 0);
+
+            }
+
+            @Override
+            public void onFailure(Call<ResponseBody> call, Throwable t) {
+                Log.d("ERROR", t.getMessage());
+            }
+        });
+    }
+
+    @Override
+    public void onConnectionSuspended(int i) {
+        client.connect();
+    }
+
+    @Override
+    public void onConnectionFailed(ConnectionResult connectionResult) {
+        BaseGameUtils.resolveConnectionFailure(this, client, connectionResult,
+                RC_SIGN_IN, R.string.signin_other_error);
     }
 }
